+++ src/main/java/be/thomasmore/qrace/repository/HighScoreRepository.java
@@ -2,5 +2,5 @@ package be.thomasmore.qrace.repository;
 import be.thomasmore.qrace.model.HighScores;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HighScoreRepository extends JpaRepository<HighScores, Long> {
+public interface HighScoreRepository extends JpaRepository<HighScores, Integer> {
 }
